tests/requirements.txt
Pillow==3.3.0
Unidecode==0.4.18
argparse==1.2.1
beautifulsoup4==4.4.1
django-admin-sortable2==0.6.4
django-classy-tags==0.7.2
django-filer==1.2.5
django-mptt==0.8.7
django-polymorphic==0.9.2
django-sekizai==0.10.0
django-treebeard==4.0.1
djangocms-helper==0.9.7
djangocms-text-ckeditor==3.3.1
easy-thumbnails==2.3
html5lib==0.9999999
jsonfield==1.0.3
pluggy==0.3.1
py==1.4.30
pytest==2.9.2
pytest-django==2.9.1
six==1.10.0
tox==2.2.1
virtualenv==13.1.2
django-reversion==1.10.2
